Only define aliases for interactive shells.

+++ common/.config/fish/config.fish
@@ -5,8 +5,6 @@ set fish_help_browser w3m
 # file permissions: rwxr-xr-x
 umask 022
 
-## environment
-
 for dir in /usr/{local,pkg}/{sbin,bin} \
 		   $HOME/Library/Python/{2.7,3.6}/bin \
 		   $HOME/.local/bin
@@ -38,7 +36,7 @@ test -d $GOPATH/bin; or mkdir -p $GOPATH/bin
 set -x PATH $PATH $GOPATH/bin
 
 if not set -q XDG_CACHE_HOME
-    set -x XDG_CACHE_HOME $HOME/.cache
+	set -x XDG_CACHE_HOME $HOME/.cache
 end
 
 # python completion
@@ -50,40 +48,42 @@ end
 # lynx style sheet
 test -e $HOME/.lynx.lss; and set -x LYNX_LSS $HOME/.lynx.lss
 
-## aliases
+if status is-interactive
+	alias m $PAGER
+	alias h "fc -l"
+	alias j jobs
+	alias ll "ls -hlaFo"
+	alias l "ls -hl"
+	alias g "egrep -l"
+	alias untar "tar xf"
+	alias md "mkdir -p"
+	alias ff "find . -name"
+	alias du "du -h"
+	alias df "df -h"
+	# paranoia
+	alias rm "rm -i"
+	# a common typo of mine
+	alias cd.. "cd .."
+	# make the ocaml repl usable
+	which ledit >/dev/null; and alias ocaml 'ledit ocaml'
+	# fun stuff
+	which curl >/dev/null; and alias weather 'curl wttr.in/dublin'
 
-alias m $PAGER
-alias h "fc -l"
-alias j jobs
-alias ll "ls -hlaFo"
-alias l "ls -hl"
-alias g "egrep -l"
-alias untar "tar xf"
-alias md "mkdir -p"
-alias ff "find . -name"
-alias du "du -h"
-alias df "df -h"
-# paranoia
-alias rm "rm -i"
-# a common typo of mine
-alias cd.. "cd .."
-# make the ocaml repl usable
-which ledit >/dev/null; and alias ocaml 'ledit ocaml'
-# fun stuff
-which curl >/dev/null; and alias weather 'curl wttr.in/dublin'
+	# csh syntax is compatible enough with fish for this to work
+	if test -x /usr/bin/dircolors
+		if test -r $HOME/.dircolours
+			eval (dircolors -c $HOME/.dircolors)
+		else
+			eval (dircolors -c)
+		end
 
-# csh syntax is compatible enough with fish for this to work
-if test -x /usr/bin/dircolors
-	test -r $HOME/.dircolours
-	and eval (dircolors -c $HOME/.dircolors)
-	or eval (dircolors -c)
-
-	alias ls "ls -F --color=auto"
-	alias grep "grep --color=auto"
-	alias fgrep "fgrep --color=auto"
-	alias egrep "egrep --color=auto"
-else
-	alias ls "ls -F"
+		alias ls "ls -F --color=auto"
+		alias grep "grep --color=auto"
+		alias fgrep "fgrep --color=auto"
+		alias egrep "egrep --color=auto"
+	else
+		alias ls "ls -F"
+	end
 end
 
 # Local config not necessarily under version control.
